# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Problem:
    properties:
      ac_count:
        description: AC 次数
        type: integer
      author:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      difficulty:
        $ref: '#/definitions/models.ProblemDifficulty'
      hint:
        type: string
      id:
        type: string
      input:
        type: string
      is_public:
        type: boolean
      memory_limit:
        description: 内存限制(MB)
        type: integer
      output:
        type: string
      sample_input:
        type: string
      sample_output:
        type: string
      source:
        type: string
      status:
        $ref: '#/definitions/models.ProblemStatus'
      submit_count:
        description: 提交次数
        type: integer
      tags:
        items:
          type: string
        type: array
      time_limit:
        description: 时间限制(ms)
        type: integer
      title:
        type: string
      updated_at:
        type: string
    required:
    - description
    - title
    type: object
  models.ProblemDifficulty:
    enum:
    - easy
    - medium
    - hard
    type: string
    x-enum-comments:
      DifficultyEasy: 简单
      DifficultyHard: 困难
      DifficultyMedium: 中等
    x-enum-descriptions:
    - 简单
    - 中等
    - 困难
    x-enum-varnames:
    - DifficultyEasy
    - DifficultyMedium
    - DifficultyHard
  models.ProblemStatus:
    enum:
    - draft
    - published
    - archived
    type: string
    x-enum-comments:
      StatusArchived: 已归档
      StatusDraft: 草稿
      StatusPublished: 已发布
    x-enum-descriptions:
    - 草稿
    - 已发布
    - 已归档
    x-enum-varnames:
    - StatusDraft
    - StatusPublished
    - StatusArchived
...
  service.RunCodeRequest:
    properties:
      code:
        type: string
      input:
        description: 自定义输入（可选）
        type: string
      language:
        type: string
    required:
    - code
    - language
    type: object
host: localhost:8080
info:
  contact:
    email: support@zkcodearena.com
    name: API Support
    url: http://localhost/support
  description: |-
    ZK Code Arena 在线编程平台 REST API 文档
    提供用户管理、题库管理、班级课程管理、代码评测等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://localhost/terms/
  title: ZK Code Arena API
  version: "1.0"
paths:
...
  /problem:
    get:
      consumes:
      - application/json
      description: 分页获取题目列表，支持按难度、标签筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 难度
        enum:
        - easy
        - medium
        - hard
        in: query
        name: difficulty
        type: string
      - collectionFormat: csv
        description: 标签列表
        in: query
        items:
          type: string
        name: tags
        type: array
      - default: false
        description: 包含私有题目（需教师或管理员权限）
        in: query
        name: include_private
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: 题目列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties: true
            type: object
      summary: 获取题目列表
      tags:
      - 题目
    post:
      consumes:
      - application/json
      description: 创建新题目（仅管理员和教师）
      parameters:
      - description: 题目信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Problem'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/models.Problem'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 创建失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建题目（教师/管理员）
      tags:
      - 题目
  /problem/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定题目（仅管理员）
      parameters:
      - description: 题目ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的题目ID
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 删除失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除题目（管理员）
      tags:
      - 题目
    get:
      consumes:
      - application/json
      description: 根据题目ID获取完整题目信息
      parameters:
      - description: 题目ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 题目详情
          schema:
            $ref: '#/definitions/models.Problem'
        "400":
          description: 无效的题目ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 题目不存在
          schema:
            additionalProperties: true
            type: object
      summary: 获取题目详情
      tags:
      - 题目
    put:
      consumes:
      - application/json
      description: 更新题目信息（管理员或题目创建者）
      parameters:
      - description: 题目ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新的题目信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Problem'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/models.Problem'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 需要登录
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 权限不足
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 题目不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 更新失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新题目（教师/管理员）
      tags:
      - 题目
  /problem/{id}/run:
    post:
      consumes:
      - application/json
      description: 在线运行代码进行测试（非提交），有频率限制
      parameters:
      - description: 题目ID
        in: path
        name: id
        required: true
        type: string
      - description: 代码和语言
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.RunCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 运行结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 题目不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 运行失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 运行代码测试
      tags:
      - 题目
  /problem/search:
    get:
      consumes:
      - application/json
      description: 根据关键词、难度、标签搜索题目
      parameters:
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 难度
        enum:
        - easy
        - medium
        - hard
        in: query
        name: difficulty
        type: string
      - collectionFormat: csv
        description: 标签列表
        in: query
        items:
          type: string
        name: tags
        type: array
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 搜索结果
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 搜索失败
          schema:
            additionalProperties: true
            type: object
      summary: 搜索题目
      tags:
      - 题目
...
  /testcase/problem/{problem_id}:
    get:
      consumes:
      - application/json
      description: 根据题目ID获取所有测试用例
      parameters:
      - description: 题目ID
        in: path
        name: problem_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 测试用例列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的题目ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 获取失败
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取题目测试用例列表
      tags:
      - 测试用例
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"